timer-stress: Add timer stress test
Adds a timer system test that launches a configurable number of threads
sleeping for a random interval between 0 and 1ms via clock_nanosleep().
This test is intended to be run in parallel with a test like cyclictest in
order to measure the impact on the real time behavior of the system.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>